Bio of Anthony Badaloo - Accountant and Justice Campaigner
Born and bred in in a farming environment in Bob Marly’s Jamaica, of Indian extraction, Anthony had a strong calling, to be an entrepreneur and justice campaigner, inspired by Mandela and Gandhi. Loves cricket, boxing and cryptocurrency as hobbies (Won’t mention Salsa Dancing).

After dipping his toes in the IT waters as a programmer, he found great congruence as a Financial Adviser with Zurich, and travelled the world on corporate hospitality. Winning many awards, for both work and community ventures..... Cont'd     https://www.publish0x.com/the-truth-about-wealth-blockchain-law/bio-of-anthony-badaloo-accountant-and-justice-campaigner-xznynqq